# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
...
        "/contracts/upload": {
            "post": {
                "description": "Upload a single document (PDF, DOCX, TXT, JPG, PNG, TIFF) with a 10MB size limit.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Upload a document for analysis",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The document to upload",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns the ID of the uploaded document",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request if the file is missing, invalid, or too large",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contracts/upload-analyze": {
            "post": {
                "description": "Uploads a document, securely stores it, performs one-step multimodal analysis for scanned PDFs, and returns the stored document_id along with structured analysis JSON.",
                "consumes": [
                    "multipart/form-data"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Upload and analyze a contract in one step",
                "parameters": [
                    {
                        "type": "file",
                        "description": "The document to upload and analyze",
                        "name": "file",
                        "in": "formData",
                        "required": true
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Returns document_id and analysis JSON",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "400": {
                        "description": "Bad request if the file is missing, invalid, or too large",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contracts/{document_id}/elements": {
            "get": {
                "description": "Analyzes a stored document to extract detailed contract elements including parties, obligations, and terms",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Detect contract elements",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "document_id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Contract elements detection result",
                        "schema": {
                            "$ref": "#/definitions/models.ContractElementsResult"
                        }
                    },
                    "400": {
                        "description": "Bad request",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contracts/{id}": {
            "get": {
                "description": "Get a single document's metadata by its unique ID.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get a document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/models.Contract"
                        }
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single document by its unique ID. This will remove the file from storage and the record from the database.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Delete a document by ID",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    },
                    "404": {
                        "description": "Document not found",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
        "/contracts/{id}/analysis": {
            "get": {
                "description": "Retrieves the structured analysis JSON for a previously analyzed document.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "Documents"
                ],
                "summary": "Get analysis results for a document",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Document ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Analysis results",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    },
                    "404": {
                        "description": "Document not found or no analysis available",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "schema": {
                            "type": "object",
                            "additionalProperties": {
                                "type": "string"
                            }
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "models.AnalysisMilestone": {
            "type": "object",
            "properties": {
                "amount": {
                    "type": "number"
                },
                "description": {
                    "type": "string"
                },
                "percentage": {
                    "type": "number"
                }
            }
        },
        "models.AnalysisRisk": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "severity": {
                    "type": "string"
                },
                "type": {
                    "type": "string"
                }
            }
        },
...
        "models.Contract": {
            "type": "object",
            "properties": {
                "analysis": {
                    "$ref": "#/definitions/models.ContractAnalysis"
                },
                "created_at": {
                    "type": "string"
                },
                "filename": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "retention_days": {
                    "description": "Default 1 year",
                    "type": "integer"
                },
                "storage_path": {
                    "type": "string"
                },
                "user_id": {
                    "type": "string"
                }
            }
        },
        "models.ContractAnalysis": {
            "type": "object",
            "properties": {
                "buyer": {
                    "type": "string"
                },
                "buyer_address": {
                    "type": "string"
                },
                "buyer_country": {
                    "type": "string"
                },
                "currency": {
                    "type": "string"
                },
                "milestones": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisMilestone"
                    }
                },
                "risk_factors": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.AnalysisRisk"
                    }
                },
                "seller": {
                    "type": "string"
                },
                "seller_address": {
                    "type": "string"
                },
                "seller_country": {
                    "type": "string"
                },
                "total_value": {
                    "type": "number"
                }
            }
        },
        "models.ContractElementsResult": {
            "type": "object",
            "properties": {
                "confidence": {
                    "type": "number"
                },
                "obligations": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContractObligation"
                    }
                },
                "parties": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContractParty"
                    }
                },
                "terms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/models.ContractTerm"
                    }
                }
            }
        },
        "models.ContractObligation": {
            "type": "object",
            "properties": {
                "deadline": {
                    "type": "string"
                },
                "description": {
                    "type": "string"
                },
                "party": {
                    "type": "string"
                },
                "type": {
                    "description": "e.g., \"payment\", \"delivery\", \"performance\"",
                    "type": "string"
                }
            }
        },
        "models.ContractParty": {
            "type": "object",
            "properties": {
                "address": {
                    "type": "string"
                },
                "contact": {
                    "type": "string"
                },
                "name": {
                    "type": "string"
                },
                "role": {
                    "description": "e.g., \"buyer\", \"seller\", \"contractor\", \"client\"",
                    "type": "string"
                }
            }
        },
...
        "models.ContractTerm": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string"
                },
                "type": {
                    "description": "e.g., \"payment\", \"termination\", \"warranty\"",
                    "type": "string"
                },
                "value": {
                    "description": "e.g., amount, duration",
                    "type": "string"
                }
            }
        },
...
    }
}